Move ExtendedProfileViewModel into separate file
